# [Java.Interop] apply AggressiveInlining where profiler shows calls